requirements.txt
streamlit==0.88.0
python-dotenv==0.19.1
transformers==4.10.0
espnet==0.10.0
soundfile==0.10.3.post1
google-generativeai==0.1.0
